+++ R/expect-self-test.R
@@ -47,8 +47,8 @@ expect_failure <- function(expr, message = NULL, ...) {
   invisible(NULL)
 }
 
-expect_snapshot_skip <- function(x) {
-  expect_snapshot_error(x, class = "skip")
+expect_snapshot_skip <- function(x, cran = FALSE) {
+  expect_snapshot_error(x, class = "skip", cran = cran)
 }
 expect_no_skip <- function(code) {
   expect_no_condition(code, class = "skip")